backend/python/pyproject.toml
[tool.poetry]
name = "ai-engineering-interview"
version = "0.1.0"
description = ""
authors = ["Maitham <maithamdib@gmail.com>", "Kimberly <kgehrmann@revdoc.com>"]
packages = [
    {include = "services"},
    {include = "main.py"}
]

[tool.poetry.dependencies]
python = "^3.9"
vital = "^1.1.22"
fastapi = "^0.77.1"
uvicorn = "^0.17.6"
python-dotenv = "^0.20.0"
langchain = "^0.2.0"
langchain-openai = "^0.1.0"
openai = "^1.30.0"
numpy = "^1.24.0"
pandas = "^2.0.0"
scipy = "^1.10.0"
scikit-learn = "^1.3.0"
fhir-parser = "^0.1.5"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
cffi = "1.17.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
